requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.0
PyMySQL>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
bcrypt>=4.0.0
pyotp>=2.9.0
pytest>=8.0.0
